feat: adjusting test and service status-ws

+++ src/main/kotlin/com/mvp/status/domain/model/payment/listener/StatusListenerDTO.kt
@@ -2,8 +2,11 @@ package com.mvp.status.domain.model.payment.listener
 
 
 import com.fasterxml.jackson.annotation.JsonProperty
+import java.math.BigDecimal
 
 data class StatusListenerDTO(
+//    @JsonProperty("id")
+//    var id: String,
     @JsonProperty("externalId")
     var externalId: String,
     @JsonProperty("idClient")
@@ -13,7 +16,7 @@ data class StatusListenerDTO(
     @JsonProperty("status")
     var status: String,
     @JsonProperty("totalPrice")
-    var totalPrice: Int,
+    var totalPrice: BigDecimal,
     @JsonProperty("waitingTime")
     var waitingTime: List<Int>
 )
